feat: enhance DDL handling with default attributes and improve recovery tests
- Added support for default attributes in table definitions by modifying the DDL processing logic.
- Updated recovery tests to initialize the buffer pool correctly.
- Modified catalog view tests to include default values in table creation.
- Enhanced the HashIndex access method to log insert and delete operations with WAL records.
- Introduced new JIT functions for summing absolute values in integer arrays, ensuring they match scalar implementations.
- Added benchmarks for TPC-H queries with DuckDB and included results for TPC-B with Ultrasql.

+++ crates/ultrasql-server/src/copy.rs
@@ -327,10 +327,10 @@ pub fn copy_out_response(n_columns: usize) -> BackendMessage {
 }
 
 #[must_use]
-pub fn copy_out_response_with_format(n_columns: usize, format_code: u16) -> BackendMessage {
+pub fn copy_out_response_with_format(n_columns: usize, format_code: u8) -> BackendMessage {
     BackendMessage::CopyOutResponse {
         overall_format: format_code,
-        column_formats: vec![format_code; n_columns],
+        column_formats: vec![u16::from(format_code); n_columns],
     }
 }
 
@@ -343,10 +343,10 @@ pub fn copy_in_response(n_columns: usize) -> BackendMessage {
 }
 
 #[must_use]
-pub fn copy_in_response_with_format(n_columns: usize, format_code: u16) -> BackendMessage {
+pub fn copy_in_response_with_format(n_columns: usize, format_code: u8) -> BackendMessage {
     BackendMessage::CopyInResponse {
         overall_format: format_code,
-        column_formats: vec![format_code; n_columns],
+        column_formats: vec![u16::from(format_code); n_columns],
     }
 }
